backend/requirements.txt
fastapi
uvicorn
pymupdf
numpy
transformers
faiss-cpu
yake
openai